Refactored code to be more organized. Removed Controllers, controllers will be specific to each app.

+++ Core/src/main/java/barbatos_rex1/laprivcore/meeting/domain/Duration.java
@@ -11,9 +11,9 @@ import lombok.*;
 @Getter
 public class Duration implements ValueObject {
     //In Minutes
-    private long duration;
+    private int duration;
 
-    public static Duration from(long duration) {
+    public static Duration from(int duration) {
         return new Duration(duration);
     }
